feat: разбивает на модули и упрощает код
