+++ js/util.js
@@ -1,0 +1,19 @@
+// функция случайного числа из диапазона
+const getRandomInteger = (a, b) => {
+  const lower = Math.ceil(Math.min(a, b));
+  const upper = Math.floor(Math.max(a, b));
+  const result = Math.random() * (upper - lower + 1) + lower;
+  return Math.floor(result);
+};
+
+// последовательная генерация чисел
+const getUniqueNumber = () => {
+  let value = 1;
+
+  return () => value++;
+};
+
+// случайный элемент в массиве
+const getRandomArrayElement = (elements) => elements[getRandomInteger(0, elements.length - 1)];
+
+export {getRandomInteger, getUniqueNumber, getRandomArrayElement};
